requirements.txt

django
psycopg2-binary
django-tailwind
whitenoise
gunicorn
dj-database-url
python-dotenv
pillow
djangorestframework
django-cors-headers